Implement revised RecyclerView adapter, tweak refreshing
New adapter has "endless scroll" capabilities that allow for loading of new values when approaching the limit of the current data set.
New adapter displays footer view that is displayed when at the bottom of the list while waiting for more data.
Refresh is now triggered if data set is empty and app is not currently refreshing.
Table is cleared on refresh.