Add page number support to FetchDealsTask and fix bug where image URL could be null.
FetchDealsTask can now receive a page number to allow for multi-page database tables.
FetchDealsTask no longer uses a bunch of ArrayList, just static variables that are stored to a ContentValues vector at the end of reading each deal.
Images are not required for OZ Bargain deals, check was added to ensure that at least a blank string is stored.